Implemented comments add and delete and also delete commnets when product is deleted

+++ week14/04.Thursday/backend-create_products/components/Product/index.js
@@ -1,13 +1,17 @@
-import useSWR, { mutate } from "swr";
+import useSWR from "swr";
 import { useRouter } from "next/router";
 import { ProductCard } from "./Product.styled";
 import Comments from "../Comments";
 import { StyledLink } from "../Link/Link.styled";
 import CommentForm from "../CommentForm";
+import ProductForm from "../ProductForm";
+import { useState } from "react";
+import { StyledButtonIcon } from "@/components/Button/Button.styled";
 
 export default function Product() {
   const router = useRouter();
   const { id } = router.query;
+  const [isEditProduct, setIsEditProduct] = useState(false);
 
   const { data, isLoading, mutate } = useSWR(`/api/products/${id}`);
 
@@ -19,15 +23,46 @@ export default function Product() {
     return;
   }
 
-  async function handleCommentDelete(commentId) {
+  async function handleDeleteProduct() {
+    const check = confirm("Are you sure you want to delete this product?");
+    if (check) {
+      const response = await fetch(`/api/products/${id}`, {
+        method: "DELETE",
+      });
+      if (response.ok) {
+        router.push("/");
+      }
+    }
+  }
+
+  async function handleEditProduct(event) {
+    event.preventDefault();
+    const formData = new FormData(event.target);
+    const productData = Object.fromEntries(formData);
+
     const response = await fetch(`/api/products/${id}`, {
-      method: "DELETE",
+      method: "PUT",
       headers: {
         "Content-Type": "application/json",
       },
-      body: JSON.stringify({ productId: id, commentId: commentId }),
+      body: JSON.stringify(productData),
     });
 
+    console.log(response);
+    if (response.ok) {
+      mutate();
+    }
+  }
+
+  async function handleCommentDelete(commentId) {
+    const response = await fetch(
+      `/api/reviews?reviewId=${commentId}&productId=${id}`,
+      {
+        method: "DELETE",
+      }
+    );
+
+    console.log("response : ", response);
     if (response.ok) {
       mutate();
     }
@@ -35,20 +70,51 @@ export default function Product() {
 
   return (
     <ProductCard>
+      <StyledButtonIcon
+        title="Edit a Product"
+        onClick={() => {
+          setIsEditProduct(!isEditProduct);
+        }}
+      >
+        <span role="image" aria-label="delete icon">
+          ✏️
+        </span>
+      </StyledButtonIcon>
+      &nbsp;&nbsp;&nbsp;
+      <StyledButtonIcon
+        onClick={handleDeleteProduct}
+        title="Delete a Product"
+        disabled={isEditProduct}
+      >
+        <span role="image" aria-label="delete icon">
+          ❌
+        </span>
+      </StyledButtonIcon>
       <h2>{data.name}</h2>
       <p>Description: {data.description}</p>
       <p>
         Price: {data.price} {data.currency}
       </p>
-      <CommentForm />
-      <h3>Comments</h3>
-      {data.reviews.length > 0 ? (
-        <Comments
-          reviews={data.reviews}
-          onCommentDelete={handleCommentDelete}
+      {isEditProduct && (
+        <ProductForm
+          onSubmit={handleEditProduct}
+          value={data}
+          isEditProduct={true}
         />
-      ) : (
-        <p>-----</p>
+      )}
+      {!isEditProduct && (
+        <>
+          <CommentForm />
+          <h3>Comments</h3>
+          {data.reviews.length > 0 ? (
+            <Comments
+              reviews={data.reviews}
+              onCommentDelete={handleCommentDelete}
+            />
+          ) : (
+            <p>-----</p>
+          )}
+        </>
       )}
       <StyledLink href="/">Back to all</StyledLink>
     </ProductCard>
